+++ script-patch-jdeploy-before.js
@@ -54,7 +54,7 @@ console.info('GroovyLint: Copied src/rules files into dist/rules');
 fse.copySync('lib', 'dist/lib');
 console.info('GroovyLint: Copied lib files into dist/lib');
 
-console.log('---- END PATCH JDEPLOY BEFORE ----');
+console.log('---- END PATCH JDEPLOY BEFORE ----\n');
 
 process.exit(0);