+++ src/groovy-lint-rules.js
@@ -104,7 +104,9 @@ const rulesFixPriorityOrder = [
     "SpaceAfterCatch",
     "SpaceAroundOperator",
     "SpaceAfterComma",
+    "SpaceAfterFor",
     "SpaceAfterIf",
+    "SpaceAfterSwitch",
     "UnnecessaryDefInFieldDeclaration",
     "UnnecessarySemicolon",
     "TrailingWhitespace",
@@ -114,6 +116,8 @@ const rulesFixPriorityOrder = [
     "UnusedImport",
     "BlockStartsWithBlankLine",
     "BlockEndsWithBlankLine",
+    "ClassStartsWithBlankLine",
+    "ClassEndsWithBlankLine",
     "MisorderedStaticImports",
     "IfStatementBraces",
     "ElseBlockBraces",
@@ -123,6 +127,10 @@ const rulesFixPriorityOrder = [
     "ConsecutiveBlankLines",
     "FileEndsWithoutNewline"
 ];
+
+// CodeNarc formatting fix rules are triggered after CodeNarc returnes violations
+// Non-CodeNarc formatting fix rules (existing only in npm-groovy-lint) must be runned always
+const formatRulesToAlwaysRun = ["IndentationClosingBraces", "IndentationComments"];
 
 const RULES_FOLDER = __dirname + "/rules";
 
@@ -153,4 +161,8 @@ function getNpmGroovyLintRules(optns = { loadTests: false }) {
     return npmGroovyLintRules;
 }
 
-module.exports = { getNpmGroovyLintRules };
+function getFormattingRulesToAlwaysRun() {
+    return formatRulesToAlwaysRun;
+}
+
+module.exports = { getNpmGroovyLintRules, getFormattingRulesToAlwaysRun };